Ajustement Api - Implémentation Svelte
Tentative d'appel de l'api depuis l'app
