--- BerryFileAPI/src/main/java/com/pamarg/fileapi/repo/ProductRepo.java
@@ -1,7 +1,0 @@
-package com.pamarg.fileapi.repo;
-
-import com.pamarg.fileapi.model.Product;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ProductRepo extends JpaRepository<Product, Long> {
-}
